feat: generalize Node.clone
