+++ src/expression.ts
@@ -87,13 +87,6 @@ export class ArrowFunctionExpr<
     this.ensure(body, "body", [NodeKind.BlockStmt]);
     this.ensureArrayOf(parameters, "parameters", [NodeKind.ParameterDecl]);
   }
-
-  public clone(): this {
-    return new ArrowFunctionExpr(
-      this.parameters.map((p) => p.clone()),
-      this.body.clone()
-    ) as this;
-  }
 }
 
 export class FunctionExpr<
@@ -109,14 +102,6 @@ export class FunctionExpr<
     this.ensure(name, "name", ["undefined", "string"]);
     this.ensureArrayOf(parameters, "parameters", [NodeKind.ParameterDecl]);
     this.ensure(body, "body", [NodeKind.BlockStmt]);
-  }
-
-  public clone(): this {
-    return new FunctionExpr(
-      this.name,
-      this.parameters.map((p) => p.clone()),
-      this.body.clone()
-    ) as this;
   }
 }
 
@@ -135,13 +120,6 @@ export class ClassExpr<C extends AnyClass = AnyClass> extends BaseExpr<
     this.ensure(heritage, "heritage", ["undefined", "Expr"]);
     this.ensureArrayOf(members, "members", ClassMember.Kinds);
   }
-  public clone(): this {
-    return new ClassExpr(
-      this.name,
-      this.heritage?.clone(),
-      this.members.map((m) => m.clone())
-    ) as this;
-  }
 }
 
 export class ReferenceExpr<
@@ -151,10 +129,6 @@ export class ReferenceExpr<
     super(NodeKind.ReferenceExpr, arguments);
     this.ensure(name, "name", ["undefined", "string"]);
     this.ensure(ref, "ref", ["function"]);
-  }
-
-  public clone(): this {
-    return new ReferenceExpr(this.name, this.ref) as this;
   }
 }
 
@@ -166,10 +140,6 @@ export class Identifier extends BaseExpr<NodeKind.Identifier> {
     this.ensure(name, "name", ["string"]);
   }
 
-  public clone(): this {
-    return new Identifier(this.name) as this;
-  }
-
   public lookup(): Decl | undefined {
     return this.getLexicalScope().get(this.name);
   }
@@ -179,10 +149,6 @@ export class PrivateIdentifier extends BaseExpr<NodeKind.PrivateIdentifier> {
   constructor(readonly name: `#${string}`) {
     super(NodeKind.PrivateIdentifier, arguments);
     this.ensure(name, "name", ["string"]);
-  }
-
-  public clone(): this {
-    return new PrivateIdentifier(this.name) as this;
   }
 
   public lookup(): Decl | undefined {
@@ -200,14 +166,6 @@ export class PropAccessExpr extends BaseExpr<NodeKind.PropAccessExpr> {
     this.ensure(expr, "expr", ["Expr"]);
     this.ensure(name, "ref", [NodeKind.Identifier, NodeKind.PrivateIdentifier]);
   }
-
-  public clone(): this {
-    return new PropAccessExpr(
-      this.expr.clone(),
-      this.name.clone(),
-      this.isOptional
-    ) as this;
-  }
 }
 
 export class ElementAccessExpr extends BaseExpr<NodeKind.ElementAccessExpr> {
@@ -215,13 +173,6 @@ export class ElementAccessExpr extends BaseExpr<NodeKind.ElementAccessExpr> {
     super(NodeKind.ElementAccessExpr, arguments);
     this.ensure(expr, "expr", ["Expr"]);
     this.ensure(element, "element", ["Expr"]);
-  }
-
-  public clone(): this {
-    return new ElementAccessExpr(
-      this.expr.clone(),
-      this.element.clone()
-    ) as this;
   }
 }
 
@@ -243,13 +194,6 @@ export class CallExpr extends BaseExpr<NodeKind.CallExpr> {
   ) {
     super(NodeKind.CallExpr, arguments);
   }
-
-  public clone(): this {
-    return new CallExpr(
-      this.expr.clone(),
-      this.args.map((arg) => arg.clone())
-    ) as this;
-  }
 }
 
 export class NewExpr extends BaseExpr<NodeKind.NewExpr> {
@@ -257,13 +201,6 @@ export class NewExpr extends BaseExpr<NodeKind.NewExpr> {
     super(NodeKind.NewExpr, arguments);
     this.ensure(expr, "expr", ["Expr"]);
     this.ensureArrayOf(args, "args", [NodeKind.Argument]);
-  }
-
-  public clone(): this {
-    return new NewExpr(
-      this.expr.clone(),
-      this.args.map((arg) => arg.clone())
-    ) as this;
   }
 }
 
@@ -273,14 +210,6 @@ export class ConditionExpr extends BaseExpr<NodeKind.ConditionExpr> {
     this.ensure(when, "when", ["Expr"]);
     this.ensure(then, "then", ["Expr"]);
     this.ensure(_else, "else", ["Expr"]);
-  }
-
-  public clone(): this {
-    return new ConditionExpr(
-      this.when.clone(),
-      this.then.clone(),
-      this._else.clone()
-    ) as this;
   }
 }
 
@@ -308,14 +237,6 @@ export class BinaryExpr extends BaseExpr<NodeKind.BinaryExpr> {
     this.ensure(left, "left", ["Expr"]);
     this.ensure(right, "right", ["Expr"]);
   }
-
-  public clone(): this {
-    return new BinaryExpr(
-      this.left.clone(),
-      this.op,
-      this.right.clone()
-    ) as this;
-  }
 }
 
 export type PostfixUnaryOp = "--" | "++";
@@ -326,20 +247,12 @@ export class UnaryExpr extends BaseExpr<NodeKind.UnaryExpr> {
     super(NodeKind.UnaryExpr, arguments);
     this.ensure(expr, "expr", ["Expr"]);
   }
-
-  public clone(): this {
-    return new UnaryExpr(this.op, this.expr.clone()) as this;
-  }
 }
 
 export class PostfixUnaryExpr extends BaseExpr<NodeKind.PostfixUnaryExpr> {
   constructor(readonly op: PostfixUnaryOp, readonly expr: Expr) {
     super(NodeKind.PostfixUnaryExpr, arguments);
     this.ensure(expr, "expr", ["Expr"]);
-  }
-
-  public clone(): this {
-    return new PostfixUnaryExpr(this.op, this.expr.clone()) as this;
   }
 }
 
@@ -350,10 +263,6 @@ export class NullLiteralExpr extends BaseExpr<NodeKind.NullLiteralExpr> {
   constructor() {
     super(NodeKind.NullLiteralExpr, arguments);
   }
-
-  public clone(): this {
-    return new NullLiteralExpr() as this;
-  }
 }
 
 export class UndefinedLiteralExpr extends BaseExpr<NodeKind.UndefinedLiteralExpr> {
@@ -362,20 +271,12 @@ export class UndefinedLiteralExpr extends BaseExpr<NodeKind.UndefinedLiteralExpr
   constructor() {
     super(NodeKind.UndefinedLiteralExpr, arguments);
   }
-
-  public clone(): this {
-    return new UndefinedLiteralExpr() as this;
-  }
 }
 
 export class BooleanLiteralExpr extends BaseExpr<NodeKind.BooleanLiteralExpr> {
   constructor(readonly value: boolean) {
     super(NodeKind.BooleanLiteralExpr, arguments);
     this.ensure(value, "value", ["boolean"]);
-  }
-
-  public clone(): this {
-    return new BooleanLiteralExpr(this.value) as this;
   }
 }
 
@@ -384,20 +285,12 @@ export class BigIntExpr extends BaseExpr<NodeKind.BigIntExpr> {
     super(NodeKind.BigIntExpr, arguments);
     this.ensure(value, "value", ["bigint"]);
   }
-
-  public clone(): this {
-    return new BigIntExpr(this.value) as this;
-  }
 }
 
 export class NumberLiteralExpr extends BaseExpr<NodeKind.NumberLiteralExpr> {
   constructor(readonly value: number) {
     super(NodeKind.NumberLiteralExpr, arguments);
     this.ensure(value, "value", ["number"]);
-  }
-
-  public clone(): this {
-    return new NumberLiteralExpr(this.value) as this;
   }
 }
 
@@ -406,20 +299,12 @@ export class StringLiteralExpr extends BaseExpr<NodeKind.StringLiteralExpr> {
     super(NodeKind.StringLiteralExpr, arguments);
     this.ensure(value, "value", ["string"]);
   }
-
-  public clone(): this {
-    return new StringLiteralExpr(this.value) as this;
-  }
 }
 
 export class ArrayLiteralExpr extends BaseExpr<NodeKind.ArrayLiteralExpr> {
   constructor(readonly items: Expr[]) {
     super(NodeKind.ArrayLiteralExpr, arguments);
     this.ensureArrayOf(items, "items", ["Expr"]);
-  }
-
-  public clone(): this {
-    return new ArrayLiteralExpr(this.items.map((item) => item.clone())) as this;
   }
 }
 
@@ -444,12 +329,6 @@ export class ObjectLiteralExpr extends BaseExpr<NodeKind.ObjectLiteralExpr> {
   constructor(readonly properties: ObjectElementExpr[]) {
     super(NodeKind.ObjectLiteralExpr, arguments);
     this.ensureArrayOf(properties, "properties", ObjectElementExpr.Kinds);
-  }
-
-  public clone(): this {
-    return new ObjectLiteralExpr(
-      this.properties.map((prop) => prop.clone())
-    ) as this;
   }
 
   public getProperty(name: string) {
@@ -496,10 +375,6 @@ export class PropAssignExpr extends BaseExpr<
     super(NodeKind.PropAssignExpr, arguments);
     this.ensure(expr, "expr", ["Expr"]);
   }
-
-  public clone(): this {
-    return new PropAssignExpr(this.name.clone(), this.expr.clone()) as this;
-  }
 }
 
 export class ComputedPropertyNameExpr extends BaseExpr<
@@ -509,10 +384,6 @@ export class ComputedPropertyNameExpr extends BaseExpr<
   constructor(readonly expr: Expr) {
     super(NodeKind.ComputedPropertyNameExpr, arguments);
     this.ensure(expr, "expr", ["Expr"]);
-  }
-
-  public clone(): this {
-    return new ComputedPropertyNameExpr(this.expr.clone()) as this;
   }
 }
 
@@ -524,10 +395,6 @@ export class SpreadAssignExpr extends BaseExpr<
     super(NodeKind.SpreadAssignExpr, arguments);
     this.ensure(expr, "expr", ["Expr"]);
   }
-
-  public clone(): this {
-    return new SpreadAssignExpr(this.expr.clone()) as this;
-  }
 }
 
 export class SpreadElementExpr extends BaseExpr<
@@ -537,10 +404,6 @@ export class SpreadElementExpr extends BaseExpr<
   constructor(readonly expr: Expr) {
     super(NodeKind.SpreadElementExpr, arguments);
     this.ensure(expr, "expr", ["Expr"]);
-  }
-
-  public clone(): this {
-    return new SpreadElementExpr(this.expr.clone()) as this;
   }
 }
 
@@ -552,23 +415,12 @@ export class TemplateExpr extends BaseExpr<NodeKind.TemplateExpr> {
     super(NodeKind.TemplateExpr, arguments);
     this.ensureArrayOf(exprs, "expr", ["Expr"]);
   }
-
-  public clone(): this {
-    return new TemplateExpr(this.exprs.map((expr) => expr.clone())) as this;
-  }
 }
 
 export class TaggedTemplateExpr extends BaseExpr<NodeKind.TaggedTemplateExpr> {
   constructor(readonly tag: Expr, readonly exprs: Expr[]) {
     super(NodeKind.TaggedTemplateExpr, arguments);
     this.ensureArrayOf(exprs, "expr", ["Expr"]);
-  }
-
-  public clone(): this {
-    return new TaggedTemplateExpr(
-      this.tag.clone(),
-      this.exprs.map((expr) => expr.clone())
-    ) as this;
   }
 }
 
@@ -577,20 +429,12 @@ export class TypeOfExpr extends BaseExpr<NodeKind.TypeOfExpr> {
     super(NodeKind.TypeOfExpr, arguments);
     this.ensure(expr, "expr", ["Expr"]);
   }
-
-  public clone(): this {
-    return new TypeOfExpr(this.expr.clone()) as this;
-  }
 }
 
 export class AwaitExpr extends BaseExpr<NodeKind.AwaitExpr> {
   constructor(readonly expr: Expr) {
     super(NodeKind.AwaitExpr, arguments);
     this.ensure(expr, "expr", ["Expr"]);
-  }
-
-  public clone(): this {
-    return new AwaitExpr(this.expr.clone()) as this;
   }
 }
 
@@ -599,20 +443,12 @@ export class PromiseExpr extends BaseExpr<NodeKind.PromiseExpr> {
     super(NodeKind.PromiseExpr, arguments);
     this.ensure(expr, "expr", ["Expr"]);
   }
-
-  public clone(): this {
-    return new PromiseExpr(this.expr.clone()) as this;
-  }
 }
 
 export class PromiseArrayExpr extends BaseExpr<NodeKind.PromiseArrayExpr> {
   constructor(readonly expr: Expr) {
     super(NodeKind.PromiseArrayExpr, arguments);
     this.ensure(expr, "expr", ["Expr"]);
-  }
-
-  public clone(): this {
-    return new PromiseArrayExpr(this.expr.clone()) as this;
   }
 }
 
@@ -626,9 +462,6 @@ export class ThisExpr<T = any> extends BaseExpr<NodeKind.ThisExpr> {
     super(NodeKind.ThisExpr, arguments);
     this.ensure(ref, "ref", ["function"]);
   }
-  public clone(): this {
-    return new ThisExpr(this.ref) as this;
-  }
 }
 
 export class SuperKeyword extends BaseNode<NodeKind.SuperKeyword> {
@@ -640,18 +473,12 @@ export class SuperKeyword extends BaseNode<NodeKind.SuperKeyword> {
   constructor() {
     super(NodeKind.SuperKeyword, arguments);
   }
-  public clone(): this {
-    return new SuperKeyword() as this;
-  }
 }
 
 export class ImportKeyword extends BaseNode<NodeKind.ImportKeyword> {
   readonly nodeKind = "Node";
   constructor() {
     super(NodeKind.ImportKeyword, arguments);
-  }
-  public clone(): this {
-    return new ImportKeyword() as this;
   }
 }
 
@@ -670,18 +497,11 @@ export class YieldExpr extends BaseExpr<NodeKind.YieldExpr> {
     this.ensure(expr, "expr", ["undefined", "Expr"]);
     this.ensure(delegate, "delegate", ["boolean"]);
   }
-  public clone(): this {
-    return new YieldExpr(this.expr?.clone(), this.delegate) as this;
-  }
 }
 
 export class RegexExpr extends BaseExpr<NodeKind.RegexExpr> {
   constructor(readonly regex: RegExp) {
     super(NodeKind.RegexExpr, arguments);
-  }
-
-  public clone(): this {
-    return new RegexExpr(this.regex) as this;
   }
 }
 
@@ -695,9 +515,6 @@ export class VoidExpr extends BaseExpr<NodeKind.VoidExpr> {
     super(NodeKind.VoidExpr, arguments);
     this.ensure(expr, "expr", ["Expr"]);
   }
-  public clone(): this {
-    return new VoidExpr(this.expr?.clone()) as this;
-  }
 }
 
 export class DeleteExpr extends BaseExpr<NodeKind.DeleteExpr> {
@@ -708,19 +525,12 @@ export class DeleteExpr extends BaseExpr<NodeKind.DeleteExpr> {
       NodeKind.ElementAccessExpr,
     ]);
   }
-  public clone(): this {
-    return new DeleteExpr(this.expr?.clone()) as this;
-  }
 }
 
 export class ParenthesizedExpr extends BaseExpr<NodeKind.ParenthesizedExpr> {
   constructor(readonly expr: Expr) {
     super(NodeKind.ParenthesizedExpr, arguments);
     this.ensure(expr, "expr", ["Expr"]);
-  }
-
-  public clone(): this {
-    return new ParenthesizedExpr(this.expr.clone()) as this;
   }
 
   public unwrap(): Expr | undefined {
@@ -734,9 +544,6 @@ export class ParenthesizedExpr extends BaseExpr<NodeKind.ParenthesizedExpr> {
 export class OmittedExpr extends BaseExpr<NodeKind.OmittedExpr> {
   constructor() {
     super(NodeKind.OmittedExpr, arguments);
-  }
-  public clone(): this {
-    return new OmittedExpr() as this;
   }
 }
